refactor(models): split batch models out of llm_execution.py (T-E05-F06-001)
Move the six batch execution models (LLMBatchStatus,
LLMBatchRequestCounts, LLMBatchSubmitRequest, LLMBatchHandle,
LLMBatchResult, BatchPollResult) verbatim into a new
models/llm_batch.py, dropping llm_execution.py from 346 to 152 lines
and back under the 350-line ceiling ahead of E05-F06's remaining
tasks. No re-export shim per no_backwards_compatibility; all ~19
importers under src/ and tests/ updated to import from the new
module. Zero behavior change.
