+++ src/agentmap/models/llm_batch.py
@@ -1,0 +1,201 @@
+"""
+Data-only batch execution models for provider-native batch LLM call contracts.
+
+Split out of ``llm_execution.py`` (E05-F06) to keep that module under the
+repository's 350-line ceiling. These models are intentionally data-only — no
+business logic lives here.
+"""
+
+from dataclasses import dataclass, field
+from enum import Enum
+from typing import Any, Dict, List, Optional
+
+from agentmap.models.llm_execution import LLMExecutionError, LLMRequest, LLMUsage
+
+
+class LLMBatchStatus(str, Enum):
+    """
+    Normalized batch lifecycle status.
+
+    AgentMap-owned superset of Anthropic's processing_status values.
+    ``submitted`` and ``failed`` are AgentMap-side derivations;
+    ``in_progress``, ``canceling``, ``ended``, and ``expired`` map 1:1
+    from the provider.
+    """
+
+    SUBMITTED = "submitted"
+    IN_PROGRESS = "in_progress"
+    CANCELING = "canceling"
+    ENDED = "ended"
+    EXPIRED = "expired"
+    CANCELED = "canceled"
+    FAILED = "failed"
+
+
+@dataclass
+class LLMBatchRequestCounts:
+    """
+    Normalized snapshot of per-item outcome counts for a batch poll result.
+
+    Fields are optional because counts may not be available for all providers
+    or at all lifecycle stages.
+    """
+
+    processing: Optional[int] = None
+    succeeded: Optional[int] = None
+    errored: Optional[int] = None
+    canceled: Optional[int] = None
+    expired: Optional[int] = None
+
+
+@dataclass
+class LLMBatchSubmitRequest:
+    """
+    Caller-owned submission descriptor for a provider-native batch call.
+
+    ``requests`` must be non-empty and contain unique ``request_id`` values.
+    ``provider`` must be one of ``"anthropic"``, ``"openai"``, or ``"google"``.
+    No ``api_key`` field — credentials are injected at adapter level.
+    """
+
+    provider: str
+    model: str
+    requests: List[LLMRequest]
+    max_tokens: Optional[int] = None
+    request_options: Dict[str, Any] = field(default_factory=dict)
+
+
+@dataclass
+class LLMBatchHandle:
+    """
+    Opaque, serializable reference to a submitted provider-native batch.
+
+    Contains only AgentMap-owned types — no ``anthropic.*`` SDK objects.
+    Provides ``to_dict()`` / ``from_dict()`` for file-backed persistence
+    without carrying credentials.
+
+    ``agentmap_batch_id`` is generated by AgentMap (``amatch_<uuid>``).
+    ``provider_batch_id`` is the provider-assigned identifier.
+    ``request_id_map`` maps caller ``request_id`` values to the ``custom_id``
+    sent to the provider (sanitized for provider constraints).
+    """
+
+    agentmap_batch_id: str
+    provider_batch_id: str
+    status: LLMBatchStatus
+    provider: str
+    model: str
+    request_id_map: Dict[str, str]
+    results_url: Optional[str] = None
+    result_ref: Optional[str] = None
+    expires_at: Optional[str] = None
+    ended_at: Optional[str] = None
+    request_counts: Optional[LLMBatchRequestCounts] = None
+
+    def to_dict(self) -> Dict[str, Any]:
+        """
+        Return a plain serializable dict suitable for json.dumps().
+
+        No ``api_key`` is included. ``status`` is serialized as its string
+        value so the dict round-trips through JSON without enum awareness.
+        """
+        counts: Optional[Dict[str, Any]] = None
+        if self.request_counts is not None:
+            counts = {
+                "processing": self.request_counts.processing,
+                "succeeded": self.request_counts.succeeded,
+                "errored": self.request_counts.errored,
+                "canceled": self.request_counts.canceled,
+                "expired": self.request_counts.expired,
+            }
+        return {
+            "agentmap_batch_id": self.agentmap_batch_id,
+            "provider_batch_id": self.provider_batch_id,
+            "status": (
+                self.status.value
+                if isinstance(self.status, LLMBatchStatus)
+                else self.status
+            ),
+            "provider": self.provider,
+            "model": self.model,
+            "request_id_map": dict(self.request_id_map),
+            "results_url": self.results_url,
+            "result_ref": self.result_ref,
+            "expires_at": self.expires_at,
+            "ended_at": self.ended_at,
+            "request_counts": counts,
+        }
+
+    @classmethod
+    def from_dict(cls, data: Dict[str, Any]) -> "LLMBatchHandle":
+        """
+        Reconstruct an ``LLMBatchHandle`` from a serialized dict.
+
+        Inverse of ``to_dict()``. Restores ``status`` as ``LLMBatchStatus``.
+        """
+        counts: Optional[LLMBatchRequestCounts] = None
+        raw_counts = data.get("request_counts")
+        if raw_counts is not None:
+            counts = LLMBatchRequestCounts(
+                processing=raw_counts.get("processing"),
+                succeeded=raw_counts.get("succeeded"),
+                errored=raw_counts.get("errored"),
+                canceled=raw_counts.get("canceled"),
+                expired=raw_counts.get("expired"),
+            )
+        return cls(
+            agentmap_batch_id=data["agentmap_batch_id"],
+            provider_batch_id=data["provider_batch_id"],
+            status=LLMBatchStatus(data["status"]),
+            provider=data["provider"],
+            model=data["model"],
+            request_id_map=dict(data["request_id_map"]),
+            results_url=data.get("results_url"),
+            result_ref=data.get("result_ref"),
+            expires_at=data.get("expires_at"),
+            ended_at=data.get("ended_at"),
+            request_counts=counts,
+        )
+
+
+@dataclass
+class LLMBatchResult:
+    """
+    Per-item result from a completed provider-native batch.
+
+    ``request_id`` is the caller-provided identifier from ``LLMRequest``.
+    ``status`` is one of: ``"succeeded"``, ``"errored"``, ``"canceled"``,
+    ``"expired"``.  ``text`` and ``usage`` are populated on success;
+    ``error`` is populated on failure.
+
+    Field names mirror ``LLMResponse``. ``resolved_model`` stays optional —
+    the Gemini batch adapter legitimately omits it.
+    """
+
+    request_id: str
+    status: str  # "succeeded" | "errored" | "canceled" | "expired"
+    resolved_provider: Optional[str] = None
+    resolved_model: Optional[str] = None
+    text: Optional[str] = None
+    usage: Optional[LLMUsage] = None
+    error: Optional[LLMExecutionError] = None
+
+
+@dataclass
+class BatchPollResult:
+    """
+    Normalized poll result returned by all adapter ``poll()`` implementations.
+
+    All adapters return this dataclass so ``LLMService.poll_batch`` can read
+    ``status`` directly without performing its own provider-specific mapping.
+    ``result_ref`` carries the provider file/object reference used to fetch
+    results (e.g. OpenAI output_file_id); it is ``None`` for providers that
+    use URL-based or inline result delivery.
+    """
+
+    status: LLMBatchStatus
+    request_counts: Optional[LLMBatchRequestCounts] = None
+    result_ref: Optional[str] = None
+    results_url: Optional[str] = None
+    ended_at: Optional[str] = None
+    expires_at: Optional[str] = None
