+++ src/agentmap/services/llm/openai_batch_adapter.py
@@ -20,15 +20,13 @@ from datetime import datetime, timezone
 from typing import Any, Dict, Generator, List, Optional, Tuple
 
 from agentmap.exceptions import LLMDependencyError, LLMServiceError
-from agentmap.models.llm_execution import (
+from agentmap.models.llm_batch import (
     BatchPollResult,
     LLMBatchRequestCounts,
     LLMBatchResult,
     LLMBatchStatus,
-    LLMExecutionError,
-    LLMRequest,
-    LLMUsage,
 )
+from agentmap.models.llm_execution import LLMExecutionError, LLMRequest, LLMUsage
 from agentmap.services.llm._batch_ids import CUSTOM_ID_RE as _CUSTOM_ID_RE
 from agentmap.services.llm._batch_ids import (
     build_request_id_map as _build_request_id_map,
