+++ tests/fresh_suite/unit/models/test_llm_batch_models_f04.py
@@ -12,7 +12,7 @@ Test cases covered:
 
 import pytest
 
-from agentmap.models.llm_execution import (
+from agentmap.models.llm_batch import (
     BatchPollResult,
     LLMBatchHandle,
     LLMBatchRequestCounts,
